requirements.txt
streamlit
pandas
google-generativeai
matplotlib
openpyxl